requirements.txt
asgiref==3.7.2
Django==4.2.5
sqlparse==0.4.4
tzdata==2023.3
djangorestframework==3.14.0
pytz==2023.3.post1
setuptools==63.2.0
typing_extensions==4.8.0